+++ packages/@runlightyear/lightyear/src/base/variable.ts
@@ -1,6 +1,7 @@
 import invariant from "tiny-invariant";
 import baseRequest from "./baseRequest";
 import { getEnvName } from "../util/getEnvName";
+import { getContext } from "./context";
 
 /**
  * @internal
@@ -30,4 +31,37 @@ export async function getVariableData(name: string): Promise<VariableData> {
   return {
     value,
   };
+}
+
+/**
+ * Set a variable on the current action or webhook.
+ *
+ * @param name
+ * @param value
+ */
+export async function setVariable(name: string, value: string | null) {
+  const envName = getEnvName();
+  const context = getContext();
+
+  const { actionName, webhookName } = context;
+
+  if (actionName) {
+    return await baseRequest({
+      method: "PATCH",
+      uri: `/api/v1/envs/${envName}/actions/${actionName}/variables/${name}`,
+      data: {
+        value,
+      },
+    });
+  } else if (webhookName) {
+    return await baseRequest({
+      method: "PATCH",
+      uri: `/api/v1/envs/${envName}/webhooks/${webhookName}/variables/${name}`,
+      data: {
+        value,
+      },
+    });
+  } else {
+    throw new Error("setVariable not called from an action or webhook context");
+  }
 }