+++ packages/@runlightyear/gcal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gcal
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.20.0
+
 ## 0.2.0
 
 ### Minor Changes